Remove special logic that stops building the response classes when an error code is received.
BEFORE THE CHANGE:
When a payment was submitted and failed to succeed, an error code was returned. The field 'avsaddr' was then placed in unconsumed_data instead of in an AddressVerificationResponse object.
It is reasonable to want to inspect additional data, such as 'avsaddr' even in failure conditions, and special logic must be used to either check AddressVerificationResponse object or unconsumed_data, depending on the situation. This makes code using this library inconsistent.

AFTER THE CHANGE:
With or without a successful transaction, objects such as AddressVerificationResponse are constructed as long as the server provides fields.
Now, we only need to check for the 'avsaddr' field in one place (the AddressVerificationResponse), never in unconsumed_data.